Basic setup
Create the index.html, main.js,
and binary_search_tree.js files.

Add boilerplate to index.html.
Set the main.js as the src
in the script tag and set its
type to module so that the
file can be loaded as a module.

In binary_search_tree.js,
initialize a Tree class that
accepts an array when
instantiated.

Initialize a Node class that
accepts a value when instantiated.
Create the leftNode and rightNode
properties and set their values
to null.

In main.js, import the Tree class.