Initialize a prepArray() method that prepares an array for a tree to be built out of it.
Take in an array and
remove duplicates by using
filter() to go
through the array and using
indexOf() to keep only the
first occurence of each
element. For each element,
it checks if the index
returned by indexOf()
(which is the first
index where that element
appears) is equal to the
current index. If so, it
includes the element;
otherwise, it skips it.

Since filter() returns an
array, sort the array
returned from filter()

Return the array.

This removes duplicates
from andand sorts the
array that was passed
into it.