Implement buildTree()
Initialize a method that
builds a balanced
binary search tree from a
an array.

Take in an array, sort and
remove duplicates.
Define start and end to
be able to set a base case.
Find the middle index and
make a node out of the
element at the middle index.
Split the array into
a left part and a right part.
Recursively build the
left subtree and right subtree
of root by passing the
leftPart and rightPart
as args to buildTree()
respectively.

Return root.