Implement prettyPrint()
In binary_search_tree.js
update the naming of the
`value` property
to `data`.

In main.js, intialize
a prettyPrint
function
that takes in a the root
of a tree as the node
parameter and console
logs the tree in a
structured format.

Create a new instance
of the Tree class,
call the buildTree()
method and pass the
result to
prettyPrint().