Implement initialize()
In the Tree class,
initialize a method that
calls the buildTree()
method on the
array property of Tree()
and returns the result.

Since buildTree()
returns the
root of the tree,
initialize() returns
the root.

Call initialize()
in the constructor.

In main.js, pass an
array into the Tree class
and update node to be
the result of
initialize().