Implement insert()
First update the name
of the data property
in the Node class to
value. Update this
name in prettyPrint
in main.js

Initialize an insert()
method that takes in
a `value` and inserts
it as a new node in the
binary search tree.

Traverse through the
tree and determine
the node that this
new node will be
attached to and return
the root node if the
value already exists
in the binary search tree.

Once the node is found,
attach the new node
to it and return the
parent node.

In main.js, call insert()
with a value.