Create rootNode property.
Previously accessing the
root node had to be done
by calling the
buildTree() on the array
property in every method
that needs root node.

Create a rootNode property
and set its value to null.

Update the `intialize` method
name and all its
references to `initialize`.

In initialize(), set the
rootNode the return value
of calling buildTree() on the
to array property and
return it.

Now methods can always access
it as needed.

In main.js, prettyPrint() can
now be called on `node` and
`node2`.