Update insert()
In insert() update the
rootNode constant to
be the rootNode property
instead of the return
value (root) of the
buildTree()
on the array property.

The previous approach
was always replacing a
previously inserted node
with a new one because
it was always working
return
value (root) of the
buildTree()
on the array property.

In main.js, insert two
nodes to the tree.