Handle case when array is empty
In the constructor, set
an empty array as the
default parameter of Tree,
so that Tree can be called
without passing in an array.
Also in the constructor,
make the rootNode property
to be the result of
buildTree() if the array
property is not empty
or null.

Clean up the initialize
method to just return
the root node.

In insert(), if there is
not root node, make the
new node the root node
and return it.

In main.js, initialize
the Tree class and
insert a node.

Pass the inserted node
into prettyPrint().

Passing the result
of calling the initialize()
method into prettyPrint()
now would not print
anything to the console
because the Tree was
not initialized with an
array(its default parameter
is anempty array) and
at this point rootNode
is null.