Implement deleting a leaf node
Initialize delete() that takes
in a `value` and and handle
deleting the node with the
given value if the node
is a leaf node.

Tranverse the tree to look
for the node with `value`
and if the node is not
found, return the rootNode.

If the target node is found
and it is a leaf node
(no children), set its
parent's reference to it
to be null.