Implement deleting a node with a child
If the target node has at
least one child, get the
child (replacementNode).

If there is no parentNode
(the node to be deleted is
the rootNode) then set the
child of the targetNode
(replacementNode) as the
root and return it.

Replace the targetNode with
its child depending on the
relationship of the targetNode
with its parent.