Implement deleting node with 2 children
Set the successor as the
node on the right of the
targetNode.
The in-order successor is
the smallest node
on the right subtree of the
targetnode. Try to find the
in-order successor by starting
at the right child of the
targetNodeand tranversing
left until you can’t
go any further.

If the inorder successor has
a parent, set the target's
left node to the successor's
right node.

Else (the right child of
the target node does not
have a leftNode, then
this right child is the
in-order successor)
set the target's right
node to the successor's
right node. (Explained
in the code comments)

Replace the value of
the targetNode with the
in-order successor's value

Return the root node.

Delete a node with 2
children in the main.js