Implement levelOrder()
Initialize a levelOrder()
method that takes in a
callback, traverses the
tree and calls the callback
on each node of the
tree.
The method tranverses
the tree in level order.

Throw an error if no
callback is passed or
the argument supplied
is not a function.

Return if the tree is
empty.

Create a queue and push
the rootNode in the queue

While the queue is not
empty, shift() it
and call the callback
on the element that
was shifted (shift
returns the element
that was removed).

Push the children of
the current element
(currentNode) if they
exist, into the queue.

Log a completed message.

In main.js pass a
callback into levelOrder().
The callback simply logs
the values of the nodes.

Comment out an alternate
method of calling the
method with a callback.