Implement preOrder()
In the Tree class,
Initialize a preOrder()
method that takes in a
callback and the currentNode
and calls the callback on
each node of the tree.
The method tranverses the
tree in preorder.

Initialize the method
with a default value
for currentNode (rootNode)
Check if the callback is
a method.
Set the base case to return
if `!currentNode`.
Call the callback on
currentNode.
Call the method recursively
on the left and right nodes
of the currentNode.

The instructions were
preOrder(callback) but
preOrder(callback, currentNode)
was implemented.

Will try to implement
inOrder(currentNode).

In main.js, call preOrder()
with a callback.