Implement inOrder
In the Tree class,
initialize a private
inOrderHelper method
takes in a `callback` and
a `currentNode` and handles
the preorder tranversal by
recursion. Call the callback
on the currentNode (every node)
as a inorder traversal requires.

Initialize an inOrder method
that takes in a `callback`.
Check if the callback is
a function.
Call the preOrderHelper
method and pass in the
callback and the rootNode.

In main.js call inOrder
with a callback.