Implement postOrder
In the Tree class,
initialize a private
postOrderHelper method that
takes in a `callback` and
a `currentNode` and handles
the postorder tranversal by
recursion. Call the callback
on the currentNode (every node)
as a postorder traversal requires.

Initialize a postOrder method
that takes in a `callback`.
Check if the callback is
a function.
Call the postOrderHelper
method and pass in the
callback and the rootNode.

In main.js call postOrder
with a callback.