Implement height()
On the Tree class,
Initialize a method, that
takes in a node and
returns its height. (The
number of edges in the
longest path, from that node
to a leaf node)

Have the base case to return
-1 if the node is null.
This is because a tree
containing only a single
node (the root) has a
height of 0. Therefore,
when a node has no children,
its subtree is considered
empty, and the height is -1.

Recursively determine the
leftHeight and rightHeight
of each subtree of the node.
Return the largest + 1.
Add 1 to account for the
node itself.

In main.js, call the
height method on a
node.