Implement depth()
In the Tree class,
initialize a method that
takes in a node and find
the number of edges in the
path from the node to the
root node.

Check if the node exists
and whether it is indeed
found in the tree and
return accordingly.

Tranverse the tree and
keep track of the edges
in the path, until
the node that is being
sought for is found.

In main.js, find the
depth of a node.