Implement isBalanced()
In Tree class
Initialize a private
\_isBalancedHelper
method that takes in a
node and checks if
all the nodes in its
subtrees are balanced.

Set the base case to
return true if !node
(the null left and
right of a leaf node)

Find the height of
the left and right
node of the the given
node. Check if the
the absoulte value
of the result of their
subtraction is greater
than 1. Return false
if so.

Else recursively
check if the
right and left subtrees
of the given node are
balanced. Combine
the results of these
checks and return them.

Initialize an isBalanced()
method that takes returns
of calling the
\_isBalancedHelper on
the root.

In main.js, check if the
tree is balanced.