Implement rebalance()
Initialize a rebalance
method that rebalances
an unbalanced tree.

Create an empty new array.
Check if the tree is
unbalanced and if it
is, traverse it inorder
and push all the values
of the nodes into the
empty array.

Call the buildTree on
the new array and set
the result as the root
node. Return it.